Some progress - with issues...

+++ src/main/java/com/example/demo/service/AccountService.java
@@ -1,18 +1,15 @@
 package com.example.demo.service;
 
 import com.example.demo.domain.entity.Account;
-import com.example.demo.domain.entity.Transaction;
 import com.example.demo.domain.model.AccountPatch;
 import com.example.demo.exception.AccountNotFoundException;
-import com.example.demo.exception.BalanceTooLowException;
 import com.example.demo.repository.AccountRepository;
-import com.example.demo.repository.TransactionRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.validation.Valid;
-import java.math.BigDecimal;
+import java.sql.SQLException;
 import java.util.List;
 
 @AllArgsConstructor
@@ -21,35 +18,51 @@ import java.util.List;
 public class AccountService {
 
     private final AccountRepository accountRepository;
-    private final TransactionRepository transactionRepository;
 
     public void deleteAccountById(int id) {
-        accountRepository.deleteById(id);
+        if (getAccountById(id).getAccountId() > 0) {
+            accountRepository.deleteById(id);
+        } else {
+            throw new AccountNotFoundException();
+        }
     }
 
     public List<Account> getAccounts() {
-        return accountRepository.findAll();
+        try {
+            return accountRepository.findAll();
+        } catch (EmptyResultDataAccessException exception) {
+            throw new AccountNotFoundException("No accounts available!");
+        }
     }
 
     public Account getAccountById(int id) {
-        return accountRepository.findById(id);
+        try {
+            return accountRepository.findById(id);
+        } catch (EmptyResultDataAccessException exception) {
+            throw new AccountNotFoundException("No account for this account id!");
+        }
     }
 
     public Account getAccountByAccountNumber(String accountNumber) {
         try {
             return accountRepository.findByAccountNumber(accountNumber);
         } catch (EmptyResultDataAccessException exception) {
-            throw new AccountNotFoundException("There is no account for this account number!");
+            throw new AccountNotFoundException("There is no internal account for this account number!");
         }
     }
 
-    public Account createAccount(Account account) {
-        return accountRepository.save(account);
+    public Account createAccount(Account account) throws SQLException {
+            return accountRepository.save(account);
     }
 
+//    this is for top up the account balance
     public Account updateAccountBalance(@Valid AccountPatch accountPatch) {
+        try{
         Account account = accountRepository.findById(accountPatch.getAccountId());
         account.setBalance(accountPatch.getBalance());
         return accountRepository.update(accountPatch.getAccountId(), accountPatch.getBalance());
+        } catch (EmptyResultDataAccessException exception) {
+            throw new AccountNotFoundException("No valid account details for patch!");
+        }
     }
 }
